homework10/requirements.txt
bcrypt==3.2.0
black==21.6b0
fastapi==0.65.2
fastapi-utils==0.2.1
Hypercorn==0.11.2
isort==5.8.0
mypy==0.902
passlib==1.7.4
peewee==3.14.4
pydantic==1.8.2
PyJWT==2.1.0
regex==2021.4.4
requests==2.25.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
